packages/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-nested-routers==0.93.4
django-cors-headers==4.3.1
python-socketio==5.10.0
python-engineio==4.8.0
psycopg[binary]==3.2.13
python-dotenv==1.0.0
gunicorn==21.2.0
